+++ src/snowflake/snowpark/functions.py
@@ -8536,23 +8536,3 @@ def locate(expr1: str, expr2: ColumnOrName, start_pos: int = 1) -> Column:
     _substr = lit(expr1)
     _str = _to_col_if_str(expr2, "locate")
     return builtin("charindex")(_substr, _str, lit(start_pos))
-
-
-def enable_span_record() -> None:
-    """
-    This function enable open telemetry span recording in action functions. If opentelemetry libraries exist in the environment, span will be recorded.
-
-    """
-    from snowflake.snowpark._internal import open_telemetry
-
-    open_telemetry.open_span_record_enabled = True
-
-
-def disable_span_record() -> None:
-    """
-    This function disable open telemetry span recording in action functions. No span will be recorded even if opentelemetry libraries exist in environment.
-
-    """
-    from snowflake.snowpark._internal import open_telemetry
-
-    open_telemetry.open_span_record_enabled = False
